refactored package of `FragmentContainer`
