+++ app/src/main/res/layout/activity_main.xml
@@ -25,7 +25,7 @@
       type="dev.teogor.ceres.main.MainViewModel" />
   </data>
 
-  <dev.teogor.ceres.m3.widgets.container.FragmentContainer
+  <dev.teogor.ceres.m3.FragmentContainer
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -152,5 +152,5 @@
       app:layout_constraintBottom_toBottomOf="parent"
       app:menu="@menu/bottom_navigation_menu" />
 
-  </dev.teogor.ceres.m3.widgets.container.FragmentContainer>
+  </dev.teogor.ceres.m3.FragmentContainer>
 </layout>
