+++ app/src/main/res/layout/fragment_about.xml
@@ -24,7 +24,7 @@
       type="dev.teogor.ceres.presentation.feature.about.AboutViewModel" />
   </data>
 
-  <dev.teogor.ceres.m3.widgets.container.FragmentContainer
+  <dev.teogor.ceres.m3.FragmentContainer
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingTop="6dp"
@@ -60,6 +60,6 @@
 
     </dev.teogor.ceres.m3.ContainerM3>
 
-  </dev.teogor.ceres.m3.widgets.container.FragmentContainer>
+  </dev.teogor.ceres.m3.FragmentContainer>
 
 </layout>
