+++ app/src/main/res/layout/fragment_home.xml
@@ -24,7 +24,7 @@
       type="dev.teogor.ceres.presentation.feature.home.HomeViewModel" />
   </data>
 
-  <dev.teogor.ceres.m3.widgets.container.FragmentContainer
+  <dev.teogor.ceres.m3.FragmentContainer
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:consumeTop="false">
@@ -57,6 +57,6 @@
       app:background_m3="primaryContainer"
       app:corner_radius="20dp"/>
 
-  </dev.teogor.ceres.m3.widgets.container.FragmentContainer>
+  </dev.teogor.ceres.m3.FragmentContainer>
 
 </layout>
