+++ ceres-m3/api/ceres-m3.api
@@ -227,6 +227,23 @@ public final class dev/teogor/ceres/m3/ErrorTextViewM3 : com/google/android/mate
 	public fun rippleMask (Z)Landroid/graphics/drawable/Drawable;
 }
 
+public final class dev/teogor/ceres/m3/FragmentContainer : androidx/constraintlayout/widget/ConstraintLayout, dev/teogor/ceres/m3/theme/ThemeHandler {
+	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;)V
+	public fun circleDrawable (I)Landroid/graphics/drawable/Drawable;
+	public fun drawable (IZ)Landroid/graphics/drawable/Drawable;
+	public fun getBackgroundColorM3 (Landroid/view/View;Ldev/teogor/ceres/m3/elevation/SurfaceLevel;)I
+	public fun getBackgroundDrawable (FLdev/teogor/ceres/m3/ColorM3;FLdev/teogor/ceres/m3/ColorM3;ZLdev/teogor/ceres/components/view/Shape;)Landroid/graphics/drawable/Drawable;
+	public fun getColorM3 (Ldev/teogor/ceres/m3/ColorM3;)I
+	public fun getMaterialShapeDrawable (Ldev/teogor/ceres/m3/elevation/SurfaceLevel;FLandroid/view/View;)Ldev/teogor/ceres/m3/drawable/ShapeDrawableM3;
+	public fun getRippleDrawable (Landroid/content/res/ColorStateList;Landroid/graphics/drawable/Drawable;Landroid/graphics/drawable/Drawable;)Landroid/graphics/drawable/RippleDrawable;
+	public fun getSchemeColor ()Ldev/teogor/ceres/m3/theme/ColorScheme;
+	public fun getSurfaceColorM3 (Ldev/teogor/ceres/m3/elevation/SurfaceLevel;)I
+	public fun log (Ljava/lang/String;)V
+	public fun onThemeChanged ()V
+	public fun rippleColorStatesList ()Landroid/content/res/ColorStateList;
+	public fun rippleMask (Z)Landroid/graphics/drawable/Drawable;
+}
+
 public final class dev/teogor/ceres/m3/ImageComponentM3 : dev/teogor/ceres/m3/ContainerBaseM3 {
 	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;)V
 	public final fun getOnClickedChangeListener ()Ldev/teogor/ceres/m3/ImageComponentM3$OnClickedChangeListener;
@@ -2359,23 +2376,6 @@ public final class dev/teogor/ceres/m3/widgets/colorpicker/view/swatch/SwatchVie
 	public synthetic fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;IILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public final fun setSwatchColor (Ldev/teogor/ceres/m3/widgets/colorpicker/model/Color;)V
 	public final fun setSwatchPatternTint (I)V
-}
-
-public final class dev/teogor/ceres/m3/widgets/container/FragmentContainer : androidx/constraintlayout/widget/ConstraintLayout, dev/teogor/ceres/m3/theme/ThemeHandler {
-	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;)V
-	public fun circleDrawable (I)Landroid/graphics/drawable/Drawable;
-	public fun drawable (IZ)Landroid/graphics/drawable/Drawable;
-	public fun getBackgroundColorM3 (Landroid/view/View;Ldev/teogor/ceres/m3/elevation/SurfaceLevel;)I
-	public fun getBackgroundDrawable (FLdev/teogor/ceres/m3/ColorM3;FLdev/teogor/ceres/m3/ColorM3;ZLdev/teogor/ceres/components/view/Shape;)Landroid/graphics/drawable/Drawable;
-	public fun getColorM3 (Ldev/teogor/ceres/m3/ColorM3;)I
-	public fun getMaterialShapeDrawable (Ldev/teogor/ceres/m3/elevation/SurfaceLevel;FLandroid/view/View;)Ldev/teogor/ceres/m3/drawable/ShapeDrawableM3;
-	public fun getRippleDrawable (Landroid/content/res/ColorStateList;Landroid/graphics/drawable/Drawable;Landroid/graphics/drawable/Drawable;)Landroid/graphics/drawable/RippleDrawable;
-	public fun getSchemeColor ()Ldev/teogor/ceres/m3/theme/ColorScheme;
-	public fun getSurfaceColorM3 (Ldev/teogor/ceres/m3/elevation/SurfaceLevel;)I
-	public fun log (Ljava/lang/String;)V
-	public fun onThemeChanged ()V
-	public fun rippleColorStatesList ()Landroid/content/res/ColorStateList;
-	public fun rippleMask (Z)Landroid/graphics/drawable/Drawable;
 }
 
 public class dev/teogor/ceres/m3/widgets/dialog/DialogBuilderM3 : androidx/appcompat/app/AlertDialog$Builder {
